feat: split page into components, wrapper in suspense to prevent vercel check errors
